Split backend implementations into modules
